+++ src/proofnetwork/xjdf/data/document.cpp
@@ -163,13 +163,14 @@ void Document::setResourceSets(const QVector<ResourceSetSP> &arg)
     }
 }
 
-void Document::addResourceSet(const ResourceSetSP &arg)
+ResourceSetSP Document::addResourceSet(const ResourceSetSP &arg)
 {
     Q_D(Document);
     auto newResourceSet = d->document.toStrongRef()->createNode<ResourceSet>();
     newResourceSet->updateFrom(arg);
     d->resourceSets << newResourceSet;
     emit resourceSetsChanged(d->resourceSets);
+    return newResourceSet;
 }
 
 DocumentSP Document::create()